Improve alignment of nav cards
