+++ src/ui/design-system/src/lib/Components/NewsCarousel/index.tsx
@@ -58,12 +58,12 @@ export const NewsCarousel: React.FC = () => {
   };
 
   return (
-    <div className="relative">
+    <div className="relative pl-8">
       {/* Navigation Buttons */}
-      <div className="absolute -left-14 top-1/2 -translate-y-1/2 flex flex-col gap-4">
+      <div className="absolute -left-6 top-1/2 -translate-y-1/2 flex flex-col gap-4">
         <button
           onClick={nextSlide}
-          className="w-10 h-10 rounded-full bg-gray-800 text-white flex items-center justify-center hover:bg-gray-700"
+          className="w-10 h-10 rounded-full bg-gray-800 text-white flex items-center justify-center hover:bg-gray-700 cursor-pointer"
           aria-label="Next slide"
         >
           <svg
@@ -83,7 +83,7 @@ export const NewsCarousel: React.FC = () => {
         </button>
         <button
           onClick={prevSlide}
-          className="w-10 h-10 rounded-full bg-gray-800 text-white flex items-center justify-center hover:bg-gray-700"
+          className="w-10 h-10 rounded-full bg-gray-800 text-white flex items-center justify-center hover:bg-gray-700 cursor-pointer"
           aria-label="Previous slide"
         >
           <svg
